+++ lib/td_bg/business_concepts/links.ex
@@ -15,14 +15,16 @@ defmodule TdBg.BusinessConcepts.Links do
     LinkCache.delete(id)
   end
 
-  def get_links(%BusinessConceptVersion{business_concept_id: business_concept_id}) do
-    get_links(business_concept_id)
+  def get_links(concept, opts \\ [])
+
+  def get_links(%BusinessConceptVersion{business_concept_id: business_concept_id}, opts) do
+    get_links(business_concept_id, opts)
   end
 
-  def get_links(%BusinessConcept{id: id}), do: get_links(id)
+  def get_links(%BusinessConcept{id: id}, opts), do: get_links(id, opts)
 
-  def get_links(business_concept_id) when is_integer(business_concept_id) do
-    {:ok, links} = LinkCache.list("business_concept", business_concept_id)
+  def get_links(business_concept_id, opts) when is_integer(business_concept_id) do
+    {:ok, links} = LinkCache.list("business_concept", business_concept_id, opts)
     links
   end
 end